Refactor porfolio to portfolio_analysis
