+++ .pre-commit-config.yaml
@@ -21,7 +21,7 @@ repos:
     hooks:
       - id: codespell
         entry: codespell
-        args: [ '--ignore-words-list="mape","ba","hist"', '--quiet-level=2', '--skip=".git,*.yaml,*.json,./tests"' ]
+        args: [ '--ignore-words-list=mape,ba,hist', '--quiet-level=2', '--skip=".git,*.yaml,*.json,./tests"' ]
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: 'v0.812'
     hooks:
