--- gamestonk_terminal/portfolio/custom_portfolio/portfolio.csv
@@ -1,6 +1,0 @@
-﻿Ticker,Shares,Other
-AAPL,5,
-GME,5,lol
-MSFT,3,
-O,2,s
-BAC,3,
