+++ gamestonk_terminal/portfolio_analysis/custom_model.py
