+++ gamestonk_terminal/portfolio_analysis/custom_portfolio/portfolio.csv
@@ -1,0 +1,6 @@
+﻿Ticker,Shares,Other
+AAPL,5,
+GME,5,lol
+MSFT,3,
+O,2,s
+BAC,3,
