+++ gamestonk_terminal/portfolio_analysis/degiro_model.py
