+++ gamestonk_terminal/portfolio_analysis/portfolio_helpers.py
